Convert to 32-bit floats

+++ src/main/kotlin/kannoo/core/CostFunction.kt
@@ -3,6 +3,6 @@ package kannoo.core
 import kannoo.math.Vector
 
 interface CostFunction {
-    fun compute(target: Vector, actual: Vector): Double
+    fun compute(target: Vector, actual: Vector): Float
     fun derivative(target: Vector, actual: Vector): Vector
 }
