+++ src/main/kotlin/kannoo/impl/SGD.kt
@@ -10,7 +10,7 @@ import kannoo.math.sumOfVector
 class SGD(
     model: Model,
     cost: CostFunction,
-    private val learningRate: Double,
+    private val learningRate: Float,
 ) {
     private val backPropagator = BackPropagator(model, cost)
